+++ src/main/java/org/ddolib/ddo/implem/solver/SequentialSolver.java
@@ -85,7 +85,7 @@ public final class SequentialSolver<T, K> implements Solver {
     /**
      * Value of the best known lower bound.
      */
-    private int bestLB;
+    private double bestLB;
     /**
      * If set, this keeps the info about the best solution so far.
      */
@@ -154,7 +154,7 @@ public final class SequentialSolver<T, K> implements Solver {
             queueMaxSize = Math.max(queueMaxSize, frontier.size());
             // 1. RESTRICTION
             SubProblem<T> sub = frontier.pop();
-            int nodeUB = sub.getUpperBound();
+            double nodeUB = sub.getUpperBound();
 
             if (verbosityLevel >= 2)
                 System.out.println("subProblem(ub:" + nodeUB + " val:" + sub.getValue() + " depth:" + sub.getPath().size() + " fastUpperBound:" + (nodeUB - sub.getValue()) + "):" + sub.getState());
@@ -208,7 +208,7 @@ public final class SequentialSolver<T, K> implements Solver {
     }
 
     @Override
-    public Optional<Integer> bestValue() {
+    public Optional<Double> bestValue() {
         if (bestSol.isPresent()) {
             return Optional.of(bestLB);
         } else {
@@ -238,7 +238,7 @@ public final class SequentialSolver<T, K> implements Solver {
      * bounds.
      */
     private void maybeUpdateBest(int verbosityLevel) {
-        Optional<Integer> ddval = mdd.bestValue();
+        Optional<Double> ddval = mdd.bestValue();
         if (ddval.isPresent() && ddval.get() > bestLB) {
             bestLB = ddval.get();
             bestSol = mdd.bestSolution();
